Add files via upload
Added a foreign key for teams to actually have Departments, so when you go in to it using /admin you can see on a team that department is -- for now this is the case as There is not departments that exist. However this substantially useful as when you want to assign a team a department you can assign it inside the team table.